fix confusing and nonsensical wording
the whole sentence is incorrect and doesn't reflect the state of big sur on these older macs. users of macs older than 2012 may read this sentence and blindly assume they'll have the superior experience, and 2012 and 2013 mac users may assume the opposite. 2012 and 2013 macs run the best with the patched operating system, not the other way around.